modify: #15 이메일 인증 담당 서버 교체
Service -> Holder로 교체
