+++ BE/service/src/config/datasource.ts
@@ -1,7 +1,6 @@
 import { DataSource } from 'typeorm';
 import { ConfigService } from '@nestjs/config';
 import { UserEntity } from 'src/entity/user.entity';
-import { EmailCodeEntity } from 'src/entity/email-code.entity';
 
 const configService: ConfigService = new ConfigService();
 
@@ -14,7 +13,7 @@ export default new DataSource({
   password: '1111',
   database: 'db',
   // synchronize: true,
-  entities: [UserEntity, EmailCodeEntity],
+  entities: [UserEntity],
   migrations: ['src/migrations/*.ts'],
   charset: 'utf8mb4_unicode_ci',
   synchronize: false,
