--- BE/service/src/entity/email-code.entity.ts
@@ -1,26 +1,0 @@
-import {
-  Entity,
-  Column,
-  CreateDateColumn,
-  PrimaryGeneratedColumn,
-} from 'typeorm';
-
-@Entity({ name: 'email-code' })
-export class EmailCodeEntity {
-  @PrimaryGeneratedColumn({ name: 'pk' })
-  pk: string;
-
-  @Column({ type: 'varchar', name: 'email' })
-  email: string;
-
-  @Column({ type: 'varchar', name: 'code' })
-  code: string;
-
-  @CreateDateColumn({
-    type: 'timestamp',
-    precision: 0,
-    name: 'created_at',
-    default: () => 'CURRENT_TIMESTAMP',
-  })
-  created_at: Date;
-}
