+++ BE/service/src/service/service-api.module.ts
@@ -4,12 +4,11 @@ import { ServiceAPIService } from './service-api.service';
 import { HttpModule } from '@nestjs/axios';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserEntity } from 'src/entity/user.entity';
-import { EmailCodeEntity } from 'src/entity/email-code.entity';
 
 @Module({
   imports: [
     HttpModule,
-    TypeOrmModule.forFeature([UserEntity, EmailCodeEntity]),
+    TypeOrmModule.forFeature([UserEntity]),
   ],
   controllers: [ServiceAPIController],
   providers: [ServiceAPIService],
