+++ BE/service/src/service/service-api.service.ts
@@ -11,8 +11,6 @@ import * as path from 'path';
 import { v4 as uuidv4 } from 'uuid';
 import { RegisterUserDto } from 'src/dto/user-register.dto';
 import { EmailCodeEntity } from 'src/entity/email-code.entity';
-import { EmailSendCodeDto } from 'src/dto/email-send-code.dto';
-import { CustomErrorException } from 'src/filter/custom-error.exception';
 import { LoginUserDto } from 'src/dto/user-login.dto';
 
 @Injectable()
@@ -50,37 +48,6 @@ export class ServiceAPIService {
     return { statusCode: 200, data: { pk: uuid } };
   }
 
-  async sendEmailCode(dto: EmailSendCodeDto) {
-    const { email } = dto;
-    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-    if (!emailRegex.test(email)) {
-      return { statusCode: 400, data: { message: 'Invalid Email' } };
-    }
-    const code = Math.floor(Math.random() * 900000) + 100000;
-
-    // TODO: Email 전송 라이브러리 사용 필요
-
-    await this.emailCodeRepository.save({
-      email: dto.email,
-      code: code.toString(),
-    });
-    return { statusCode: 200 };
-  }
-
-  async verfiyEmailCode(email: string, code: string) {
-    const emailRow = await this.emailCodeRepository.findOne({
-      where: { email: email },
-    });
-    if (!emailRow) {
-      return { result: false, message: 'Email Not Exist' }
-    }
-    const isCodeMatch = emailRow.code === code;
-    if (!isCodeMatch) {
-      return { result: false, message: 'Code is not match' }
-    }
-    return { result: true };
-  }
-
   async loginUser(dto: LoginUserDto) {
     const { id, password } = dto;
     const userRow = await this.userRepository.findOne({ where: { id } });
@@ -92,18 +59,6 @@ export class ServiceAPIService {
       return { statusCode: 400, data: { message: 'Password is not match' } };
     }
     return { statusCode: 200 };
-  }
-
-  // Issuer에게 학생 email - 학번 매칭 여부 검증
-  async verifyMajorMatch(email: string, studentNumber: string) {
-    // TODO: test.env, launch.env에 추가
-    // /api/issuer/verify-match
-    const url = this.configService.get<string>('API_VERIFY_MAJOR_MATCH');
-    return lastValueFrom(
-      this.httpService
-        .get(url, { params: { email, studentNumber } })
-        .pipe(map((response) => response?.data)),
-    );
   }
 
   /*
